Rename RingBuffer.hpp to ring_buffer.hpp
Update include paths in test files to match the new filename.
